Delete customer endpoint fix

+++ src/main/java/com/webtp/gimme/service/CustomerService.java
@@ -11,11 +11,14 @@ import com.webtp.gimme.repository.CustomerRepository;
 import java.util.List;
 
 import com.webtp.gimme.security.CustomerDetails;
+import com.webtp.gimme.security.CustomerDetailsService;
 import org.apache.logging.log4j.util.Strings;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -24,6 +27,9 @@ public class CustomerService {
 
     @Autowired
     private CustomerRepository customerRepository;
+
+    @Autowired
+    private CustomerDetailsService customerDetailsService;
 
     @Autowired
     private PasswordEncoder passwordEncoder;
@@ -44,8 +50,13 @@ public class CustomerService {
         return 200;
     }
 
-    public void deleteCustomer(String username) {
+    public void deleteCustomer(String username, Authentication authentication) {
+        CustomerDetails customerDetails = (CustomerDetails) authentication.getPrincipal();
+        if (!customerDetails.getCustomer().getUsername().equals(username)) {
+            throw new AccessDeniedException("Vous n'avez pas les droits nécessaires pour supprimer cet utilisateur.");
+        }
         customerRepository.deleteById(username);
+        customerDetailsService.evictCache(username);
     }
 
     public void updateCustomer(Customer user) {
